+++ src/LIAnsureProtect.Domain/Common/IDomainEvent.cs
@@ -1,6 +1,0 @@
-namespace LIAnsureProtect.Domain.Common;
-
-public interface IDomainEvent
-{
-    DateTime OccurredAtUtc { get; }
-}