+++ src/LIAnsureProtect.Domain/Common/IHasDomainEvents.cs
@@ -1,8 +1,0 @@
-namespace LIAnsureProtect.Domain.Common;
-
-public interface IHasDomainEvents
-{
-    IReadOnlyCollection<IDomainEvent> DomainEvents { get; }
-
-    void ClearDomainEvents();
-}